requirements.txt
transformers
datasets
tqdm
pytorch_lightning
mlflow
memray
